requirements.txt
numpy
pandas
opencv-python
pillow
tqdm
pyyaml
matplotlib
scikit-learn

torch
torchvision